+++ src/source/UI/NewUI/Combat/NewUIDuelWatchUserListWindow.cpp
@@ -89,8 +89,11 @@ bool CNewUIDuelWatchUserListWindow::Render()
     float fFontHeight = TextSize.cy / g_fScreenRate_y;
 
     POINT ptSize = {57, 17};
-    POINT ptOrigin = {m_Pos.x, m_Pos.y - (ptSize.y + 1) * (long)g_DuelMgr.GetDuelWatchUserCount() +
-                                   (ptSize.y - (long)fFontHeight) / 2 + 1};
+    POINT ptOrigin = {
+        static_cast<LONG>(m_Pos.x),
+        static_cast<LONG>(m_Pos.y - (ptSize.y + 1) * static_cast<long>(g_DuelMgr.GetDuelWatchUserCount()) +
+                          (ptSize.y - static_cast<long>(fFontHeight)) / 2 + 1)
+    };
 
     for (int i = 0; i < g_DuelMgr.GetDuelWatchUserCount(); ++i)
     {
