+++ src/source/UI/NewUI/Inventory/NewUIMixInventory.cpp
@@ -50,7 +50,10 @@ bool CNewUIMixInventory::Create(CNewUIManager* pNewUIMng, int x, int y)
 
     LoadImages();
 
-    POINT ptBtn = { m_Pos.x + (long)INVENTORY_WIDTH * (long)0.5f - (long)22.f, m_Pos.y + (long)380 };
+    POINT ptBtn = {
+        static_cast<LONG>(m_Pos.x + INVENTORY_WIDTH * 0.5f - 22.f),
+        static_cast<LONG>(m_Pos.y + 380),
+    };
 
     m_BtnMix.ChangeButtonImgState(true, IMAGE_MIXINVENTORY_MIXBTN, false);
     m_BtnMix.ChangeButtonInfo(m_Pos.x + INVENTORY_WIDTH * 0.5f - 22.f, m_Pos.y + 380, 44.f, 35.f);
